+++ packages/svelte-query/tests/context/context.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest'
-import { getIsRestoringContext } from '../src/context'
+import { getIsRestoringContext } from '../../src/context'
 
 describe('getIsRestoringContext', () => {
-  it('Should not throw when called outside of a component', async () => {
+  it('Should not throw when called outside of a component', () => {
     expect(() => getIsRestoringContext()).to.not.throw()
   })
 })
