+++ packages/svelte-query/tests/createMutation/CreateMutation.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { QueryClient } from '@tanstack/query-core'
-  import { setQueryClientContext } from '../src/context'
-  import { createMutation } from '../src/createMutation'
-  import type { CreateMutationOptions } from '../src/types'
+  import { setQueryClientContext } from '../../src/context'
+  import { createMutation } from '../../src/createMutation'
+  import type { CreateMutationOptions } from '../../src/types'
 
   export let options: CreateMutationOptions
