+++ packages/svelte-query/tests/createQueries/CreateQueries.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import { createQueries } from '../src/createQueries'
+  import { createQueries } from '../../src/createQueries'
   import type { QueryClient } from '@tanstack/query-core'
-  import type { QueriesOptions, QueriesResults } from '../src/createQueries'
+  import type { QueriesOptions, QueriesResults } from '../../src/createQueries'
 
   export let options: {
     queries: [...QueriesOptions<any>]
